+++ CHANGES.rst
@@ -2,6 +2,19 @@
 CHANGES
 =======
 
+----------
+0.3.0-dev0
+----------
+
+* Implemented a ZookeeperMasterDetector.  Just pass a 'zk://ensemble/path'
+  URI to the PesosSchedulerDriver and it will automatically be selected.  Must
+  use 'pesos[zk]' requirement for this support.  By default 'pesos' can only
+  talk to explicit PIDs.
+
+* Implements reliable scheduler registration.
+
+* Add basic test coverage for the scheduler.
+
 -----
 0.2.2
 -----